validation.go file with Validate functions
