+++ x/did/client/cli/txDidDocument.go
@@ -4,8 +4,6 @@ import (
 	"encoding/json"
 	"io/ioutil"
 
-	//"time"
-
 	"github.com/spf13/cast"
 	"github.com/spf13/cobra"
